# API/swagger.yaml
swagger: "2.0"
info:
  title: Car Rental API
  version: 1.0.0
  description: A simple API for managing car rentals and user authentication.
host: localhost:3000
basePath: /api
schemes:
  - http
paths:
...
  /user/delete/{id}:
    delete:
      summary: Delete a rental booking
      description: Deletes a rental booking by its ID if the user has the proper access rights.
      parameters:
        - in: path
          name: id
          required: true
          type: integer
          description: The ID of the rental booking to delete
      responses:
        '200':
          description: Rental booking successfully canceled
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Rental booking successfully canceled"
        '403':
          description: Forbidden access
          schema:
            type: object
            properties:
              error:
                type: string
                example: "No access to cancel this rental booking"
        '404':
          description: Rental booking or car not found
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Rental booking or car not found"
        '500':
          description: Internal Server Error
          schema:
            type: object
            properties:
              error:
                type: string
                example: "An error occurred during the cancellation process"